Added radius of gyration calculation and explanation for HPC job submission
I edited the vmd_analysis.tcl file to include code to calculate the radius of gyration from a simulation, which necessitated the inclusion of the center_of_mass,tcl and gyr_radius.tcl files. I also updated the README.md file to include instructions for running an analysis on a supercomputer.